rename project to 'CredentialsManager' because you can not have class and namespace with same name.
