--- Credentials/Credentials.cs
@@ -1,6 +1,0 @@
-﻿namespace Credentials
-{
-  public class Credentials
-  {
-  }
-}
