+++ CredentialsManager/Credentials.cs
@@ -1,0 +1,7 @@
+﻿namespace CredentialsManager
+{
+  public static class Credentials
+  {
+    private static readonly string CredentialsFile = "Users.txt";
+  }
+}
