+++ CredentialsManager/CredentialsManager.csproj
