+++ CredentialsTests/CredentialsTests.csproj
@@ -27,10 +27,6 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Credentials\Credentials.csproj"/>
-    </ItemGroup>
-
-    <ItemGroup>
         <None Update="Input\read1.txt">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
@@ -49,6 +45,10 @@
         <None Update="Output\expected_multi_line.txt">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\CredentialsManager\CredentialsManager.csproj"/>
     </ItemGroup>
 
 </Project>
